pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agents-with-intent"
version = "0.1.0"
description = "A Python library for building LangGraph-based AI agents with progressive skill discovery following the Agent Skills specification"
readme = "README.md"
authors = [
    {name = "Tomer Gal"}
]
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/Tomer-Gal/Agents-With-Intent"
Repository = "https://github.com/Tomer-Gal/Agents-With-Intent"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents_with_intent*"]
